pyproject.toml
[project]
name = "news-mvp"
version = "0.1.0"
description = "Unified news MVP with ETL"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "pyyaml>=6.0",
  "structlog>=24.1",
  "click>=8.1",
  "rich>=13.7",
  "python-dotenv>=1.0",
  "pandas>=2.2",
  "pyarrow>=17.0",
  "requests>=2.32",
  "feedparser>=6.0",
  "beautifulsoup4>=4.12",
  "lxml>=5",
  "polars>=1.6",
  "httpx>=0.27",
  "duckdb>=1.0",
  "pillow>=10.0"
]

[project.scripts]
news-mvp = "news_mvp.cli:app"

[project.optional-dependencies]
validation = ["great_expectations>=0.18"]
dev = [
  "pre-commit>=3.6",
  "pytest>=8.0",
  "ruff>=0.3.0",
  "pyright>=1.1"
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
